Fix trendingScore type

+++ Sources/HFAPI/Hub/Dataset.swift
@@ -54,7 +54,7 @@ public struct Dataset: Identifiable, Codable, Sendable {
     public let downloadsAllTime: Int?
 
     /// The trending score.
-    public let trendingScore: Int?
+    public let trendingScore: Double?
 
     /// The used storage in bytes.
     public let usedStorage: Int?
